requirements.txt
streamlit
matplotlib
python-docx
google-genai
pandas
openpyxl
streamlit-authenticator==0.1.5
gspread
oauth2client
numpy
requests
google-api-python-client
google-generativeai
streamlit-authenticator

